Homepage: move css & js around
CSS should be at the top and JS should be at the bottom.
This is best practice since you want the styling rules to be on the page
before the html body content gets inserted.

Since we only use the JS after the html is loaded, we can put all the JS
at the bottom which lets the body render faster.

Also use the minified version of jQuery from cdnjs (saves a lot in size)